python-microservice/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
pydantic==2.12.2
pydantic_core==2.41.4
psycopg2-binary==2.9.11
python-dotenv==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
